third_party/blink/web_tests/svg/dynamic-updates/SVGCircleElement-svgdom-cy-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'cy' property of the SVGCircleElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS circleElement.cy.baseVal.value is -50
PASS circleElement.cy.baseVal.value is 150
PASS successfullyParsed is true

TEST COMPLETE
